requirements.txt
yfinance 
nsetools 
pandas
mysql-connector-python

